// packages/ui/src/core/tokens/generatedColors.ts
export const backgroundColors = {
  backgroundDefaultDefault: 'var(--background-default-default)',
  backgroundDefaultDefaultHover: 'var(--background-default-default-hover)',
  backgroundDefaultSecondary: 'var(--background-default-secondary)',
  backgroundDefaultSecondaryHover: 'var(--background-default-secondary-hover)',
  backgroundDefaultTertiary: 'var(--background-default-tertiary)',
  backgroundDefaultTertiaryHover: 'var(--background-default-tertiary-hover)',
  backgroundDisabledDefault: 'var(--background-disabled-default)',
  backgroundDangerDefault: 'var(--background-danger-default)',
  backgroundDangerHover: 'var(--background-danger-hover)',
  backgroundDangerSecondary: 'var(--background-danger-secondary)',
  backgroundDangerSecondaryHover: 'var(--background-danger-secondary-hover)',
  backgroundBrandDefault: 'var(--background-brand-default)',
  backgroundBrandDefaultHover: 'var(--background-brand-default-hover)',
  backgroundBrandSecondary: 'var(--background-brand-secondary)',
  backgroundBrandSecondaryHover: 'var(--background-brand-secondary-hover)',
  backgroundBrandTertiary: 'var(--background-brand-tertiary)',
  backgroundBrandTertiaryHover: 'var(--background-brand-tertiary-hover)',
  backgroundTagBlue: 'var(--background-tag-blue)',
  backgroundTagGreen: 'var(--background-tag-green)',
  backgroundTagYellow: 'var(--background-tag-yellow)',
  backgroundTagPurple: 'var(--background-tag-purple)',
  backgroundTagRed: 'var(--background-tag-red)',
  backgroundTagGray: 'var(--background-tag-gray)',
  backgroundStateDone: 'var(--background-state-done)',
  backgroundStateProgress: 'var(--background-state-progress)',
  backgroundStateToDo: 'var(--background-state-to-do)'
} as const

export const borderColors = {
  borderDefaultDefault: 'var(--border-default-default)',
  borderDefaultSecondary: 'var(--border-default-secondary)',
  borderDefaultTertiary: 'var(--border-default-tertiary)',
  borderDisabledDefault: 'var(--border-disabled-default)',
  borderDangerDefault: 'var(--border-danger-default)',
  borderBrandDefault: 'var(--border-brand-default)',
  borderBrandSecondary: 'var(--border-brand-secondary)',
  borderBrandTertiary: 'var(--border-brand-tertiary)'
} as const
// ...
export const textColors = {
  textDefaultDefault: 'var(--text-default-default)',
  textDefaultSecondary: 'var(--text-default-secondary)',
  textDefaultTertiary: 'var(--text-default-tertiary)',
  textDisabledDefault: 'var(--text-disabled-default)',
  textDisabledOnDisabled: 'var(--text-disabled-on-disabled)',
  textDangerDefault: 'var(--text-danger-default)',
  textDangerSecondary: 'var(--text-danger-secondary)',
  textDangerTertiary: 'var(--text-danger-tertiary)',
  textDangerOnDanger: 'var(--text-danger-on-danger)',
  textDangerOnDangerSecondary: 'var(--text-danger-on-danger-secondary)',
  textDangerOnDangerTertiary: 'var(--text-danger-on-danger-tertiary)',
  textBrandDefault: 'var(--text-brand-default)',
  textBrandSecondary: 'var(--text-brand-secondary)',
  textBrandTertiary: 'var(--text-brand-tertiary)',
  textBrandOnBrand: 'var(--text-brand-on-brand)',
  textBrandOnBrandSecondary: 'var(--text-brand-on-brand-secondary)',
  textBrandOnBrandTertiary: 'var(--text-brand-on-brand-tertiary)',
  textTagBlue: 'var(--text-tag-blue)',
  textTagGreen: 'var(--text-tag-green)',
  textTagYellow: 'var(--text-tag-yellow)',
  textTagPurple: 'var(--text-tag-purple)',
  textTagRed: 'var(--text-tag-red)',
  textTagGray: 'var(--text-tag-gray)'
} as const